requirements.txt
flask
flask-restx
python-dotenv
requests